Use time resolution.
Instead of free running the sytem clock and simulating an absurd amount
of steps in the engine, break it into discrete steps. "time_sim.h" file
is provided to change the time resolution of the engine.

Making the resolution smaller will result in increased fidelity of the
simulation at the tradeoff of performance.