Working Buffer
Engine hands clients a buffer which is activley being worked on, which
can cause visual artifacting. Maintin an internal working scratchpad and
only allow clients to access a finalzed buffer in a good state.

This buffer might be updated while the client is reading it, but all the
values should be fine to draw.