Hard-Enforce Energy Conservation
I suspect due to some impossibilities expressing all real numbers,
we have situations where energy is consistently not conserved. As
a mitigaiton, we go back in time 1 step and replay the collision
at a finer granularity. This had a positive effect (about 30% of
collisions can be fixed) but some cannot.

We bail out of energy is not conserved but this is an incremental
improvement.