Don't Latch UID When Rearranging Particles
When rearranging particles to make them fit on the screen, we were
erroneously latching the uid during copy. Due to the way the latch
is implemented this should never be done as part of a copy manually
anyway.

Since the particles hadn't been assigned a UID yet they were then
latched to the default 0 value which, on top of being invalid,
is used to index into the particle array. This caused us to access
memory offsets like size_t::max() which does not work well.