Lazy-Evaluate Vector Magnitude, Kinetic Energy
Every single time we modify a vector, we would re-calculate its
magnitude. This involves a minimum of 3x squaring operations,
3x additions, and 1x square root. Given that in many cases this
value is never used (e.g. particle travels thru space with no
interactions) it was very often re-calculated, and then never used.

Instead, make these values private and access them through a member
accessor. The accessor calculates these values on-demand and avoids
a large amount of throw-away calculations.

For example without this feature our median cycle time was:

3989us

After implementing this for vector magnitude, it was lowered to:

1846us, which is a 2.16x speed increase.

A smaller and more difficult (and arguably not worth having done
but already done) optimization for calculating kinetic energy was
implemented which further decreased median cycle time to:

1776us

A small victory for much more code change but now that the work is
done I'm happy to commit it.