Provide Vector & Particle Move Ctor / Assignment
For some completely unholy reason, despite assignment for primitives
(on which Vectors and thus Particles are based) presumably being
as expensive as copying those same primitives we see about a 7%
performance increase when providing move semantics for these objects.

Perhaps this allows the compiler to make some absurd optimization
when dealing with temporaries, further invesetigation is needed.